requirements.txt
Pillow
opencv-python-headless
imagehash
reportlab
numpy
customtkinter